+++ Example/Tests/NSURLSessionTests.swift
@@ -70,11 +70,11 @@ class NSURLSessionTests: XCTestCase {
         let URL = NSURL(string: "https://www.example.com/1")!
         let body1 = "Test response 1".dataUsingEncoding(NSUTF8StringEncoding)!
         let request1 = NSURLRequest.init(URL: URL)
-        NSURLSession.mockSingle(request1, body: body1)
+        NSURLSession.mockNext(request1, body: body1)
         
         let body2 = "Test response 2".dataUsingEncoding(NSUTF8StringEncoding)!
         let request2 = NSURLRequest.init(URL: URL)
-        NSURLSession.mockSingle(request2, body: body2)
+        NSURLSession.mockNext(request2, body: body2)
         
         // Create a session
         let conf = NSURLSessionConfiguration.defaultSessionConfiguration()
@@ -177,7 +177,7 @@ class NSURLSessionTests: XCTestCase {
         let body = "Test response 1".dataUsingEncoding(NSUTF8StringEncoding)!
         let request = NSURLRequest.init(URL: URL)
         let headers = ["Content-Type" : "application/test", "Custom-Header" : "Is custom"]
-        NSURLSession.mockSingle(request, body: body, headers: headers, statusCode: 200)
+        NSURLSession.mockNext(request, body: body, headers: headers, statusCode: 200)
         
         // Create a session
         let conf = NSURLSessionConfiguration.defaultSessionConfiguration()
@@ -285,7 +285,6 @@ class NSURLSessionTests: XCTestCase {
             XCTAssertEqual(delegate.dataKeyedByTaskIdentifier[task1.taskIdentifier], "123456".dataUsingEncoding(NSUTF8StringEncoding))
             XCTAssertEqual(delegate.dataKeyedByTaskIdentifier[task2.taskIdentifier], "654321".dataUsingEncoding(NSUTF8StringEncoding))
         }
-
     }
 
     func testSession_WithFailureBlock_ShouldReturnError() {